refactor: switch to dnd-kit's PointerSensor
